+++ R/calculate_coobs.R
@@ -51,31 +51,31 @@ calculate_coobs <- function(mraster,
   
   #--- check for required packages ---#
   if (!requireNamespace("doParallel", quietly = TRUE)) {
-    stop("Package \"doParallel\" needed for this function to work. Please install it.",
+    stop("Package \"doParallel\" is needed for this function to work. Please install it.",
          call. = FALSE
     )
   }
   
   if (!requireNamespace("doSNOW", quietly = TRUE)) {
-    stop("Packag \"doSNOW\" needed for this function to work. Please install it.",
+    stop("Packag \"doSNOW\" is needed for this function to work. Please install it.",
          call. = FALSE
     )
   }
   
   if (!requireNamespace("foreach", quietly = TRUE)) {
-    stop("Package \"foreach\" needed for this function to work. Please install it.",
+    stop("Package \"foreach\" is needed for this function to work. Please install it.",
          call. = FALSE
     )
   }
   
   if (!requireNamespace("snow", quietly = TRUE)) {
-    stop("Package \"snow\" needed for this function to work. Please install it.",
+    stop("Package \"snow\" is needed for this function to work. Please install it.",
          call. = FALSE
     )
   }
   
   if (!requireNamespace("Rfast", quietly = TRUE)) {
-    stop("Package \"Rfast\" needed for this function to work. Please install it.",
+    stop("Package \"Rfast\" is needed for this function to work. Please install it.",
          call. = FALSE
     )
   }